code/python/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
adal==1.2.4
    # via msrestazure
arrow==0.16.0
    # via
    #   -r requirements.in
    #   m26
astroid==2.4.2
    # via pylint
attrs==20.2.0
    # via pytest
azure-cognitiveservices-anomalydetector==0.3.0
    # via -r requirements.in
azure-cognitiveservices-formrecognizer==0.1.1
    # via -r requirements.in
azure-cognitiveservices-inkrecognizer==1.0.0b1
    # via -r requirements.in
azure-cognitiveservices-language-luis==0.7.0
    # via -r requirements.in
azure-cognitiveservices-speech==1.13.0
    # via -r requirements.in
azure-cognitiveservices-vision-computervision==0.6.0
    # via -r requirements.in
azure-cognitiveservices-vision-face==0.4.1
    # via -r requirements.in
azure-common==1.1.25
    # via
    #   azure-cognitiveservices-anomalydetector
    #   azure-cognitiveservices-formrecognizer
    #   azure-cognitiveservices-language-luis
    #   azure-cognitiveservices-vision-computervision
    #   azure-cognitiveservices-vision-face
azure-core==1.8.1
    # via
    #   azure-cognitiveservices-inkrecognizer
    #   azure-cosmos
    #   azure-eventhub
    #   azure-storage-blob
    #   azure-storage-file-datalake
azure-cosmos==4.0.0
    # via -r requirements.in
azure-eventhub==5.2.0
    # via -r requirements.in
azure-iot-device==2.2.0
    # via -r requirements.in
azure-storage-blob==12.4.0
    # via
    #   -r requirements.in
    #   azure-storage-file-datalake
azure-storage-file-datalake==12.1.2
    # via -r requirements.in
certifi==2020.6.20
    # via
    #   matplotlib
    #   msrest
    #   requests
    #   uamqp
cffi==1.14.3
    # via cryptography
chardet==3.0.4
    # via
    #   pdfminer.six
    #   requests
click==7.1.2
    # via nltk
coverage==5.3
    # via pytest-cov
cryptography==3.1.1
    # via
    #   adal
    #   azure-storage-blob
    #   pdfminer.six
cycler==0.10.0
    # via matplotlib
dask[dataframe]==2.27.0
    # via -r requirements.in
docopt==0.6.2
    # via -r requirements.in
et-xmlfile==1.0.1
    # via openpyxl
flake8==3.8.3
    # via -r requirements.in
fsspec==0.8.3
    # via dask
idna==2.10
    # via requests
iniconfig==1.0.1
    # via pytest
isodate==0.6.0
    # via msrest
isort==5.5.3
    # via pylint
janus==0.4.0
    # via azure-iot-device
jdcal==1.4.1
    # via openpyxl
jinja2==2.11.2
    # via -r requirements.in
joblib==0.16.0
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.2.0
    # via matplotlib
lazy-object-proxy==1.4.3
    # via astroid
locket==0.2.0
    # via partd
m26==0.2.1
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.2
    # via -r requirements.in
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
more-itertools==8.5.0
    # via pytest
msrest==0.6.19
    # via
    #   azure-cognitiveservices-anomalydetector
    #   azure-cognitiveservices-formrecognizer
    #   azure-cognitiveservices-language-luis
    #   azure-cognitiveservices-vision-computervision
    #   azure-cognitiveservices-vision-face
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   msrestazure
msrestazure==0.6.4
    # via azure-cognitiveservices-language-luis
mysql-connector-python==8.0.21
    # via -r requirements.in
nltk==3.5
    # via -r requirements.in
numpy==1.19.2
    # via
    #   -r requirements.in
    #   dask
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==3.0.5
    # via -r requirements.in
packaging==20.4
    # via pytest
paho-mqtt==1.5.1
    # via azure-iot-device
pandas==1.1.2
    # via
    #   -r requirements.in
    #   dask
partd==1.1.0
    # via dask
pdfminer.six==20200726
    # via -r requirements.in
pillow==7.2.0
    # via matplotlib
pluggy==0.13.1
    # via pytest
protobuf==3.13.0
    # via mysql-connector-python
psutil==5.7.2
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.9.0
    # via pytest
pyarrow==1.0.1
    # via -r requirements.in
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pyjwt==1.7.1
    # via adal
pylint==2.6.0
    # via -r requirements.in
pymongo==3.11.0
    # via -r requirements.in
pyodbc==4.0.30
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pysocks==1.7.1
    # via azure-iot-device
pytest-cov==2.10.1
    # via -r requirements.in
pytest==6.0.2
    # via
    #   -r requirements.in
    #   pytest-cov
python-dateutil==2.8.1
    # via
    #   adal
    #   arrow
    #   matplotlib
    #   pandas
pytz==2020.1
    # via pandas
pyyaml==5.3.1
    # via dask
redis==3.5.3
    # via -r requirements.in
regex==2020.7.14
    # via nltk
requests-oauthlib==1.3.0
    # via msrest
requests-unixsocket==0.2.0
    # via azure-iot-device
requests==2.24.0
    # via
    #   -r requirements.in
    #   adal
    #   azure-core
    #   azure-iot-device
    #   msrest
    #   requests-oauthlib
    #   requests-unixsocket
scikit-learn==0.23.2
    # via -r requirements.in
scipy==1.5.2
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.15.0
    # via
    #   astroid
    #   azure-core
    #   azure-cosmos
    #   azure-iot-device
    #   cryptography
    #   cycler
    #   isodate
    #   msrestazure
    #   packaging
    #   protobuf
    #   python-dateutil
    #   uamqp
sortedcontainers==2.2.2
    # via pdfminer.six
sqlalchemy==1.3.19
    # via -r requirements.in
threadpoolctl==2.1.0
    # via scikit-learn
toml==0.10.1
    # via
    #   pylint
    #   pytest
toolz==0.11.1
    # via
    #   dask
    #   partd
tqdm==4.49.0
    # via
    #   -r requirements.in
    #   nltk
uamqp==1.2.10
    # via azure-eventhub
urllib3==1.25.10
    # via
    #   azure-iot-device
    #   requests
    #   requests-unixsocket
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
